Use integer division everywhere.
This commit changes the behavior of `_update_icon`! Previously a
character in `blocks` was chosen by rounding a division up to the
nearest integer. As a result, if a discharging battery was even 1%
charged the second character in `blocks` would be chosen. Now instead of
rounding up to the nearest integer we will round to the nearest integer
(rounding up in the event of a tie).