Add threshold parameters for colors.
This commit changes the behavior of `_set_bar_text`! If `hide_when_full`
is True, the response text will be set to '' if the battery level  is
greater than or equal to the new parameter `threshold_full`, which has
default value 99 (not 100).

This commit changes the behavior of `_set_bar_color`! The response color
will be set to the 'good' color when the battery level is considered
neither bad nor degraded. (Previously, this would only happen when the
battery level was 100.) In addition, the boundaries between 'bad' and
'degraded' and 'degraded' and 'good' are now handled slightly
differently: by default, the battery will have to exceed 10 to be
considered degraded, and it will have to exceed 30 to be considered
good. These values are controlled with the new parameters
`threshold_bad` and `threshold_degraded`.