New module: coin_balance
The 'coin_balance' module fetches the users balances of different
crypto-currencies from a user-defined server. The currently supported
coins are: Bitcoin, Dogecoin, and Litecoin. 'coin_balance' requires a
server that is compatible with the Bitcoin RPC protocol and is easily
extendable to new crypto-currencies by extending the 'COIN_PORTS'
dictionary.