coin_balance: Add config support for credetials
This patch adds the ability to configure the credentials for the
selected coins using the '<coin>_username' and '<coin>_password'
settings. <coin> has to be replaced by any of the currenctly supported
identifiers (bitcoin, dogecoin, litecoin).

In addition to the coin specific credentials there is also support for
'global' credentials via the 'username' and 'password' settings. These
credentials will be used for coins for which no specific credentials are
defined. If neither the 'global' nor the 'specific' credentials are
configured, 'coin_balance' will try to acquire the them from the
respective daemon's configuration of each activated coin.