Implement battery measurement using /sys interface
It works, only that I get a vastly different time remaining compared to
acpi. However at least it seems to not be completely off.

And of course, this new method of measurement is much more safe and does
not have any dependency.

Only tested with one battery. However if the old code for measuring
multiple batteries works it should work too.