do_not_disturb: Avoid usage of killall
This commit removes usage of killall in do_not_disturb module to fix
this module under NixOS. In NixOS "killall dunst" returns a "dunst:
process not found" (this happens with other packages too, thanks to the
usage of binary wrappers in NixOS).

Instead of "killall", we use "pkill", since it is more reliable and is
available everywhere.