Repeat connection attempt more often if issue caused by network (owm)
If DNS is not available during a period of time when OWM does a request
(typically during boot until the DHCP lease is obtained via WiFi), user
has to wait cache_timeout for the next connection attempt.  This commit
forces an update every 10 seconds for failures caused by network issue.