implement global request_retry_times and request_retry_wait options
Requests failing due to network unavailability or remote server timeouts are
retried automatically ``request_retry_times`` times (default ``3``) at a
``request_retry_wait`` (default ``2``) seconds interval.

This allows to be more graceful to i3 startup when network is not up yet or to
short network disruptions and not display an error on the bar in that case.

This is related to the MR #1842 inspired by the work of @jose1711