imap module: guard against possible race condition in IDLE thread
The IDLE thread indicates that a change has occurred (and so new mail
should be checked for) by refreshing the module and ending itself.

Because on refresh the module checks for the existence of the IDLE
thread to see if mail should be checked, there is a chance that the
module is already refreshing before the thread has completely finished,
and therefore the mail would not be checked.

To guard against this an attribute idle_thread_ending is set before
refreshing the module with the IDLE thread.