mpd_status using idle
Issuing commands every two seconds, even if nothing changes causes high
CPU load with Mopidy-MPD, and possibly other MPD servers. With polling
every 2 seconds (the default), mopidy (while paused) was consuming
between 10% and 20% cpu; using idling, this number dropped to 0.0%.

This patch is similar to IDLE in imap.py, where we listen for changes on
subsystems we're interested in (99% that's just 'player', but we'll
default to 'mixer' (changes in volume) and 'options' (e.g. repeat mode)
too, just in case a user has those states displayed in her status bar).