Support more battery in battery_level module
Previous code only matches with "BAT*" names in the /sys/class/power_supply.
There does not seem to be any naming convention in the documentation:
https://www.kernel.org/doc/html/latest/power/power_supply_class.html
For example, the battery on the GPD Pocket is exposed as "max170xx_battery".

This patch allows for easier addition of glob patterns to use to match for
battery nodes and adds "*bat*" in the list.