Set up standard logging module
This has the advantage of overall less custom code; as well as support
for per-module configuration. This would enable a potential solution for
https://github.com/ultrabug/py3status/issues/1479, since in the future
it can allow per-module configuration of log levels.

I expect this to mainly help module creators, allowing them to enable
logging for only their module, while disabling all other messages. It
also is easier to use, since the `logging` module's interface is
generally simpler than `self.py3`.

Here, I've made the decision to keep the message format as close as
possible to the existing log messages.