Use git package for logging git commit information if available
The reporting of an underlying git source does not work in a range of
instances.  For example, if the repo is part of a submodule, then no git
information will be reported.

Using the python git module gives more consistent information about what
git repo that we are running from.  Try and import the git package and
fall back to reading the files manually if it is unavailable.