volume_status module: retry backend init. on load
The volume_status module must properly handle the case where the module is
loaded before the audio backend (Pipewire, PulseAudio, etc.) is initialized.
This was raised in issue #2136.

PR #2139 addresses this by adding a `start_delay` property. This is not ideal,
as it needs to be set/tuned by the user, and is fragile (a delay sufficient for
the worst case will usually cause the module to load slowly in the usual case).

Instead, retry initialize the audio backend with a sleep between tries.
In total, we can sleep up to ~30 seconds which ought to be enough for anybody.