playerctl: Improve caching algorithm
In order for the module to update continuously, a couple of things need
to happen:
1. The {position} placeholder needs to be used
2. At least one player needs to be playing AND have a valid position

Also, the status strings were changed to represent the defaults provided
by playerctl instead of doing a .lower() call.

The cache_timeout is exposed as a valid configuration option, but it is
hidden to give the users the impression that the module's updates are
done automatically and not on an interval.