Remove default title escape in playerctl module
I originall saw the need for escaping titles because I had pango set to
markup in my bar config. However, removing that removes the need for
escaping. If users still use pango markup in their bar config, they can
add the escape themselves in the format_player. Therefore, this is
potentially a breaking change for some users.

This specifically was found when the escape translated an ampersand (&)
to &amp;, which then showed up in the title. I also had a `max_length`
in the format_player. When the max length cut off right in the middle of
the escaped sequence, the bar would not show any output.